DTS Integration Complete

+++ Source/SmartInventory/SmartInventory/Controllers/LibraryController.cs
@@ -2678,8 +2678,8 @@ namespace SmartInventory.Controllers
                 fillRegionProvinceDetail(objPOD, GeometryType.Point.ToString(), geom);
                 //Fill Parent detail...              
                 fillParentDetail(objPOD, new NetworkCodeIn() { eType = EntityType.POD.ToString(), gType = GeometryType.Point.ToString(), eGeom = objPOD.geom }, networkIdType);
-                objPOD.longitude = Convert.ToDecimal(geom.Split(' ')[0]);
-                objPOD.latitude = Convert.ToDecimal(geom.Split(' ')[1]);
+                objPOD.longitude = Convert.ToDouble(geom.Split(' ')[0]);
+                objPOD.latitude = Convert.ToDouble(geom.Split(' ')[1]);
                 // Item template binding
                 var objItem = BLItemTemplate.Instance.GetTemplateDetail<PODTemplateMaster>(Convert.ToInt32(Session["user_id"]), EntityType.POD);
                 Utility.MiscHelper.CopyMatchingProperties(objItem, objPOD);
